fixed foreign id datatype

+++ src/main/java/edu/uniandes/hotelandes/hotel/room/booking/RoomBooking.java
@@ -8,4 +8,4 @@ public record RoomBooking(
     Integer hotelRoomId,
     Date checkIn,
     Date checkOut,
-    Byte consumptionPlanId) {}
+    Integer consumptionPlanId) {}
